Partial rewrite of the ClamAV signature translation script to add support for the LDB database. Also fixed a bug causing signatures targeting any filetype to be ignored. The Yara module now receives information regarding the version_info location. Wrote a simple executable to be used in unit tests.

+++ plugins/plugin_resources.cpp
@@ -58,7 +58,7 @@ public:
 			yara::const_matches matches = y.scan_bytes(*(*it)->get_raw_data());
 			if (matches->size() > 0)
 			{
-				for (int i = 0 ; i < matches->size() ; ++i)
+				for (size_t i = 0 ; i < matches->size() ; ++i)
 				{
 					std::string ext = matches->at(i)->operator[]("extension");
 					if (ext == ".exe" || ext == ".sys" || ext == ".cab")
